requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
opencensus
opencensus-ext-grpc
grpcio_reflection
